requirements.txt
fastapi
uvicorn
torch
transformers
pillow
accelerate
pydantic
